+++ src/test/scala/tscfg/generators/scala/ScalaMainSpec.scala
@@ -1175,4 +1175,40 @@ class ScalaMainSpec extends AnyWordSpec {
       assert(c.example.c.head.dddd.eeee === 30)
     }
   }
+
+  "issue 73 - Ability to extend or implement external type" when {
+    "73a @define abstract extends java.lang.Object" should {
+      "generate AbstractA extends java.lang.Object" in {
+        val r = ScalaGen.generate("example/issue73a.spec.conf")
+        assert(r.code contains "extends java.lang.Object")
+      }
+
+      "usual parsing" in {
+        val c = ScalaIssue73aCfg(ConfigFactory.parseString("""test.impl {
+            |  a = "aa"
+            |  b = "bb"
+            |}
+            |""".stripMargin))
+        assert(c.test.impl.a === "aa")
+        assert(c.test.impl.b === "bb")
+      }
+    }
+
+    "73b @define abstract implements java.io.Serializable" should {
+      "generate AbstractA extends java.io.Serializable" in {
+        val r = ScalaGen.generate("example/issue73b.spec.conf")
+        assert(r.code contains "extends java.io.Serializable")
+      }
+
+      "do usual parsing" in {
+        val c = ScalaIssue73bCfg(ConfigFactory.parseString("""test.impl {
+            |  a = "aa"
+            |  b = "bb"
+            |}
+            |""".stripMargin))
+        assert(c.test.impl.a === "aa")
+        assert(c.test.impl.b === "bb")
+      }
+    }
+  }
 }